v1.0.6: Add secure signature to gem
This mitigates some security and authenticity concerns brought
about by the recent compromise of RubyGems.org. The changes in
this update deal solely with making sure the gem code you receive
from RubyGems is the same code we pushed up there in the last
release.

As of now, all future Rain releases will be signed with a
secure key.

http://www.coffeepowered.net/2013/01/31/how-to-sign-your-ruby-gems/