v1.0.8: Write ReleaseTag to String in YAML
There was a bug introduced very recently wherein the String
representation of ReleaseTag was not being coerced out of
the object by YAML, and therefore would store ReleaseTags
in YAML as a Ruby object. This proved problematic since
we expect the config file data to be strictly of the
String persuasion.

There is also some improvements to the testing framework
so an Exception will be raised whenever you attempt to
run the test suite with uncommitted changes. On some
tests, this makes the whole suite hang and is problematic
to go and debug if you're not expecting it.

Changes:

- Write test for versions_path
- Ignore all further changes to versions.yml in test
- Fix file path references to ./config/versions.yml when testing
- Write more intricate tests for update_release_tag
- Add test for whether hash is returning strings
- Test whether YAML was updated to a String
- Always coerce ReleaseTags into a String before writing to YAML..