Ensure Bundler has a clean env before exec_cmd'ing
Execute all Git commands in a `Bundler.clean_env` to ensure we're
using the right versions of everything. Fixes an issue where previously
a `rain` deploy could not be combined with any "Git hook"-creating
plugins, such as `git-tracker`, or else the commit would fail due to
the improper Ruby being used.

Thanks to @stevenharman and @alindeman for showing us this
awesome little Bundler trick, which helps us mitigate this issue and any
future problems that may be encountered in the course of Rain's use with
other Git plugins.